requirements.txt
python-firebase
Django==2.0.2
pytz==2018.3
coala
coala-bears